requirements.txt
pyfakefs==4.6.1

pyyaml==6.0
requests